Project Start
The basic server was created to listen on a port for incoming
connections where a thread is executed to access your instance of the
back-end server.
The basic game was created to start working form.
Image Diarie will will be used as a time line of the games progress for
later demonstration of what has been achieved.