Made corrections and added use case diagrams
Changed some of the concrete use case names to verbs, that was not
verbs.
Moved showLiveLeadedBoard to inherit from game menu and not game play.
Added new game and included it in game menu.
Corrected some names to CamelCase, that was not already CamelCase.
Added four sub-level use case diagrams: Login, Sign up and Game menu.