Server
Connected the client side to the server side.
Added the following functionality server side:
To send an email with the client's password to him, if the client's
requested it.
To check client's login.
To add a new client with all his details to the database.
Retrieve a client's number of zombie kills from the database.
Store the client's number of zombie kills in the database.
Increase the client's number of zombie kills by one in the database.
Stores a client's password in the database.