Made changes to server
Client side, added all the new calls that can be made to the server.

Made the Database object a global variable instead of creating the
object in every function.

Added the functionality server side:

When password is saved in database, it gets encrypted.
When password(Password is encrypted) is retrieved, it gets decrypted.
If the user forgets his username and password, he can retrieve his
password, by giving his email address.
A OAuth user table.
Experience column to user table.
The leader board.