Finished server unit test and unique password encryption
Finished the server unit testing and is now on version 2.0
Added the password encryption to be unique for each user by using his or
her internet modem IP address to add to the password before encryption
and removing it after decryption. Also added the commons-codec-19.jar
library for the encryption and decryption. The password are also not
only now encrypted in the database, but also when sending them across
the network.