Added functionality and fixed bugs.
removed an unnecessary white space in the whatismyip url.

Added a test to serverUnitTesting to test if a connection to the server
can be made.

Fix bug in the database class where the availableUsername was returning
the wrong value.

Made it so that the error and audit logs works concurrently.